# Api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/mfa/setup/start": {
      "post": {
        "operationId": "MfaAccountController_setupStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMfaSetupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Begin opt-in MFA enrolment (password re-auth). Returns the otpauth URI + QR to import into an authenticator app. MFA is NOT active until /setup/confirm.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaSetupStartResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/setup/confirm": {
      "post": {
        "operationId": "MfaAccountController_setupConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMfaSetupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirm enrolment with the first code: activates MFA and returns the one-time backup codes (shown ONCE).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaSetupConfirmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/disable": {
      "post": {
        "operationId": "MfaAccountController_disable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaReauthDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Disable MFA (password + a current TOTP or backup code). Clears MFA state and revokes remembered devices."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/backup-codes/regenerate": {
      "post": {
        "operationId": "MfaAccountController_regenerateBackupCodes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaReauthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replace the backup codes (password + a current TOTP or backup code); returns the new set ONCE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaSetupConfirmResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/admin-reset": {
      "post": {
        "operationId": "MfaAccountController_adminReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminResetMfaDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Administrator-only: reset a target operator’s MFA and revoke their sessions, devices, and live challenges."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/devices": {
      "get": {
        "operationId": "MfaAccountController_listDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List the operator’s active trusted (\"remember this device\") devices — empty when the feature is off.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RememberedDeviceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/devices/{id}": {
      "delete": {
        "operationId": "MfaAccountController_revokeDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoke one of the operator’s own trusted devices (the next sign-in from it re-prompts for MFA)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/password/admin-reset": {
      "post": {
        "operationId": "PasswordResetAdminController_adminReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Administrator-only: set a working password for a target operator who cannot self-serve the MFA-gated reset, and revoke their sessions, devices, and live challenges. No session/token is issued. Self-reset via this path is forbidden (use the self-service reset)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/dashboard/stream-token": {
      "post": {
        "operationId": "RealtimeController_streamToken",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sets a short-lived (60s) httpOnly cookie (`ftd_stream`, scope `stream:read`) used to authenticate GET /dashboard/stream. No body — the credential is never exposed to JS."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fintech Dashboard API",
    "description": "System-of-record backend (NestJS on Fastify + Prisma + PostgreSQL).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StartMfaSetupDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "description": "The operator’s current password — re-auth for this sensitive change."
          }
        },
        "required": [
          "password"
        ]
      },
      "MfaSetupStartResponseDto": {
        "type": "object",
        "properties": {
          "otpauthUri": {
            "type": "string",
            "description": "otpauth:// URI to import into an authenticator app."
          },
          "qrDataUrl": {
            "type": "string",
            "description": "PNG data-URL of the QR encoding the otpauth URI."
          }
        },
        "required": [
          "otpauthUri",
          "qrDataUrl"
        ]
      },
      "ConfirmMfaSetupDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "example": "123456",
            "description": "The first 6-digit code from the authenticator app, proving enrolment."
          }
        },
        "required": [
          "code"
        ]
      },
      "MfaSetupConfirmResponseDto": {
        "type": "object",
        "properties": {
          "backupCodes": {
            "description": "One-time backup recovery codes — shown ONCE; store them now.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "backupCodes"
        ]
      },
      "MfaReauthDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "description": "The operator’s current password."
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "A current 6-digit TOTP code OR a one-time backup code (XXXXX-XXXXX)."
          }
        },
        "required": [
          "password",
          "code"
        ]
      },
      "AdminResetMfaDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "The target operator whose MFA enrolment is to be reset."
          }
        },
        "required": [
          "userId"
        ]
      },
      "RememberedDeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "ipPrefix": {
            "type": "string",
            "description": "Coarse network prefix bound at issue — NOT a full IP."
          }
        },
        "required": [
          "id",
          "createdAt",
          "expiresAt",
          "ipPrefix"
        ]
      },
...
      "AdminPasswordResetDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "string",
            "format": "uuid",
            "description": "The target operator whose password is to be reset."
          },
          "newPassword": {
            "type": "string",
            "minLength": 12,
            "maxLength": 64,
            "description": "The new password (min 12 chars; must include upper, lower, digit, and symbol)."
          }
        },
        "required": [
          "targetUserId",
          "newPassword"
        ]
      },
...
    }
  }
}